+++ shop-Alejandro-Samuel/src/main/java/es/iesclaradelrey/da2d1e2425/shopalejandrosamuel/services/RatingServiceImpl.java
@@ -26,7 +26,7 @@ public class RatingServiceImpl implements RatingService {
     }
 
     @Override
-    public Optional<Rating> findById(int id) {
+    public Optional<Rating> findById(Long id) {
         return this.ratingRepository.findById(id);
     }
