+++ shop-Alejandro-Samuel/src/main/java/es/iesclaradelrey/da2d1e2425/shopalejandrosamuel/services/TypeServiceImpl.java
@@ -21,7 +21,7 @@ public class TypeServiceImpl implements TypeService {
     }
 
     @Override
-    public Optional<Type> findById(int id) {
+    public Optional<Type> findById(Long id) {
         return typeRepository.findById(id);
     }
